audit: replace isMember read-gates with tryAccess; consolidate two drift items
isMember vs read-access — `StreamMemberRepository.isMember` and
`StreamService.isMember` answer "is this user a stream_members row",
not "can this user read this stream". The two are NOT equivalent:
public channels grant read to any workspace member without a row, and
threads inherit access from the root. Several handlers were gating
read-only / participation flows on plain isMember, blocking workspace
members from interacting with public-channel content they could
already see. Switched to `streamService.tryAccess` (the canonical
read-access check):

- attachments/handlers.ts download endpoint: previously a workspace
  member who could read a public channel's messages got 403 trying to
  download files attached to those messages.
- messaging/handlers.ts addReaction + removeReaction: blocked
  workspace members from reacting to messages in public streams.
- messaging/handlers.ts getHistory: blocked them from viewing edit
  history of messages they could already see.
- messaging/handlers.ts update + delete: kept the "author-only"
  restriction below tryAccess so the gate is "can read AND is the
  author" rather than "is a stream_members row AND is the author".
  Avoids blocking authors of messages in public channels they never
  formally joined.

Drift consolidation:

- escapeXmlAttr was triplicated (lib/ai/message-formatter,
  agents/quote-resolver, agents/companion/context). Hoisted to
  lib/xml.ts; all three callers import the shared helper.
- The shared-message pointer card body renderer was duplicated
  between the TipTap NodeView (`SharedMessageView`) and the markdown
  paragraph swap (`SharedMessagePointerBlock`). Extracted to a single
  `SharedMessageCardBody` component in components/shared-messages/;
  both surfaces own only their frame (NodeView wrapper / Link wrapper)
  and route the per-state body through the shared component.
