+++ apps/backend/src/lib/ai/message-formatter.ts
@@ -3,13 +3,10 @@ import type { Message } from "../../features/messaging"
 import type { AttachmentWithExtraction } from "../../features/attachments"
 import { UserRepository } from "../../features/workspaces"
 import { PersonaRepository } from "../../features/agents"
+import { escapeXmlAttr } from "../xml"
 
 function escapeXml(s: string): string {
   return s.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;")
-}
-
-function escapeXmlAttr(s: string): string {
-  return escapeXml(s).replace(/"/g, "&quot;")
 }
 
 /**
