+++ apps/frontend/src/lib/markdown/shared-message-block.tsx
@@ -1,9 +1,8 @@
 import { Share2 } from "lucide-react"
 import { Link, useParams } from "react-router-dom"
 import { cn } from "@/lib/utils"
-import { Skeleton } from "@/components/ui/skeleton"
-import { useSharedMessageSource, type SharedMessageSource } from "@/hooks/use-shared-message-source"
-import { MarkdownContent } from "@/components/ui/markdown-content"
+import { useSharedMessageSource } from "@/hooks/use-shared-message-source"
+import { SharedMessageCardBody } from "@/components/shared-messages/card-body"
 
 interface SharedMessagePointerBlockProps {
   streamId: string
@@ -17,12 +16,7 @@ interface SharedMessagePointerBlockProps {
  * passes through markdown rendering (i.e. the timeline, thread panel, and any
  * other surface that consumes `contentMarkdown`). The in-composer card lives
  * in `SharedMessageView` (a TipTap NodeView); both route through the shared
- * resolver hook so cache behavior matches.
- *
- * The body is rendered as full markdown via `MarkdownContent` — this card is
- * the live inline surfacing of the source message, not a sidebar/preview
- * snippet (INV-60 doesn't apply), so emoji shortcodes, mentions, formatting,
- * and code blocks all render as they would for the original message.
+ * resolver hook + `SharedMessageCardBody` so cache + render behavior matches.
  */
 export function SharedMessagePointerBlock({ streamId, messageId, authorName }: SharedMessagePointerBlockProps) {
   const { workspaceId } = useParams<{ workspaceId: string }>()
@@ -37,7 +31,9 @@ export function SharedMessagePointerBlock({ streamId, messageId, authorName }: S
       data-type="shared-message"
     >
       <Share2 className="mt-0.5 h-4 w-4 shrink-0 text-muted-foreground" aria-hidden="true" />
-      <div className="min-w-0 flex-1">{renderBody(authorName, source)}</div>
+      <div className="min-w-0 flex-1">
+        <SharedMessageCardBody source={source} fallbackAuthor={authorName} />
+      </div>
     </div>
   )
 
@@ -48,46 +44,4 @@ export function SharedMessagePointerBlock({ streamId, messageId, authorName }: S
       {card}
     </Link>
   )
-}
-
-function renderBody(fallbackAuthor: string, source: SharedMessageSource) {
-  if (source.status === "deleted") {
-    return (
-      <>
-        <AuthorLabel name={fallbackAuthor || "—"} />
-        <p className="mt-0.5 italic text-muted-foreground">Message deleted by author</p>
-      </>
-    )
-  }
-
-  if (source.status === "missing") {
-    return (
-      <>
-        <AuthorLabel name={fallbackAuthor || "—"} />
-        <p className="mt-0.5 italic text-muted-foreground">Message no longer available</p>
-      </>
-    )
-  }
-
-  if (source.status === "pending") {
-    return (
-      <>
-        <AuthorLabel name={fallbackAuthor || "—"} />
-        {source.showSkeleton ? <Skeleton className="mt-1 h-3 w-48" /> : <p className="mt-0.5 h-3" aria-hidden="true" />}
-      </>
-    )
-  }
-
-  return (
-    <>
-      <AuthorLabel name={source.authorName || fallbackAuthor || "—"} />
-      <div className="mt-0.5">
-        <MarkdownContent content={source.contentMarkdown} className="text-sm leading-relaxed" />
-      </div>
-    </>
-  )
-}
-
-function AuthorLabel({ name }: { name: string }) {
-  return <span className="text-xs font-medium text-foreground/80">{name}</span>
 }
